pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"] # list of folders that contain the packages (["."] by default)

[project]
name = "mypackagename" # appears on pip list only
dynamic = ["version"]
dependencies = [
    "langchain>=0.3.23",
    "langchain-ollama>=0.3.2",
    "langgraph>=0.3.34",
    "pydantic>=2.0",
    "langchain-core",
    "numpy",
    "langchain_community",
    "tracer @ git+https://github.com/AhmedYasserrr/tracer.git",
    "python-dotenv",
]

requires-python = ">=3.10,<3.13"
authors = [
  { name = "Omar Abdelgawad", email = "omar.abdelgawad@ejust.edu.eg" },
  { name = "Eyad Hussein", email = "eyad.hussein@ejust.edu.eg" },
  { name = "Ahmed Hassan", email = "Ahmed.hasan@ejust.edu.eg" },
  { name = "Ahmed Gamea", email = "Ahmed.gamea@ejust.edu.eg" },
  { name = "Abdelaziz Serour", email = "abdelaziz.serour@ejust.edu.eg" },
  # 2 are missing
]
description = "A Python package for agents that have os-level knowledge."
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["agents", "flow", "crew", "llm", "ollama", "knowledge"]
classifiers = [
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[dependency-groups]
dev = ["mypy", "ruff", "pytest", "pytest-cov", "tox", "pre-commit"]

[tool.setuptools.package-data]
os_assistant = ["py.typed"]

[project.urls]
Repository = "https://github.com/eyad-hussein/os_assistant"

###################################################################################################
[tool.pytest.ini_options]
addopts = "--cov=os_assistant" #include all packages under src directory
testpaths = ["tests"]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.ruff]
src = ["src"]
extend-exclude = ["notebooks"]

[tool.ruff.lint]
extend-select = ["UP", "I"]
